+++ Front/src/pages/TravelPlan.tsx
@@ -3,12 +3,12 @@ import * as S from "../styles/travelPlan.style";
 import { useParams } from "react-router-dom";
 import { Container, Wrapper } from "../styles/layout.style";
 import travelPlanData from "../../data/travelPlan.json";
-import citiesList from "../../data/citiesList.json";
 import scheduleList from "../../data/scheduleList.json";
 import StarsIcon from "../assets/StarsIcon";
 import { GroupButton } from "../styles/button.style";
 import BasicMap from "../components/map/BasicMap";
 import ScheduleItem from "../components/schedule/ScheduleItem";
+import { getCityMap, getCityName } from "../utils/getCityData";
 
 const TravelPlan: FC = () => {
   const { travelPlanId } = useParams();
@@ -20,32 +20,14 @@ const TravelPlan: FC = () => {
     (plan) => plan.id === Number(travelPlanId)
   );
 
-  const getCityName = (city_id: number): string => {
-    const city = citiesList.find((city) => city.id === city_id);
-    return city ? city.name : "국내";
-  };
-
-  const getCityMap = (city_id: number) => {
-    const city = citiesList.find((city) => city.id === city_id);
-    if (!city) {
-      return { lat: 35.723817, lng: 127.483131, level: 14 };
-    }
-
-    return {
-      lat: city.city_latitude ?? 35.723817,
-      lng: city.city_longitude ?? 127.483131,
-      level: city.city_level ?? 14,
-    };
-  };
   // 나중엔 /api/travel_plans/{travel_plan_id} GET 요청
 
   if (!travelPlan) {
     return <h1>not found</h1>; // TODO: error page 만들기
   }
 
-  const handleClickChevron = (id: number) => {
+  const handleClickChevron = (id: number) =>
     setOpenSchedules((prev) => ({ ...prev, [id]: !prev[id] }));
-  };
 
   const { city_id, start_date, end_date } = travelPlan;
   const { lat, lng, level } = getCityMap(city_id);
